Update index table

update_index will generate a tibble object and, if selected, display it with md formatting. It will contain the title of the analysis, as extracted from the .rmd with a link to the corresponding .html. If chosen, it will also show the number of the analysis and a description of the analysis, as extracted from the file name and .rmd, respectively.

update_index(
  html_folder = "docs",
  include_number = TRUE,
  include_description = TRUE,
  md_formatting = TRUE
)

Arguments

html_folder

Character. The folder containing .htmls to be referenced in the contents table. This path should be relative to the root folder of the R project (i.e. the folder containing the .Rproj file). Default is "docs/".

include_number

boolean. Whether to include the numbering of the analysis i.e. if the document has been labelled with a number to indicate ordering (e.g. "01_example.rmd"). If no numbering requested, analyses will be ordered alphabetically. Default is TRUE.

include_description

boolean. Whether to include a description of the analysis, which will be taken straight from the .rmd. This function assumes that that this description is described in the section preceded by "> Aim:" and ended with "<br><br>", as in the example analysis templates available in both templates. Both of these templates were generated using: https://github.com/RHReynolds/rmdplate. Default is TRUE.

md_formatting

boolean. Whether tibble should be returned with md formatting. Default is TRUE.

Value

A tibble, containing the title of each analysis, as extracted from the .rmd in the user-specified folder, with a link to the corresponding .html. It will optionally also contain the number of the analysis and a description. By default, it will be returned with md formatting, but this can be set to FALSE.